Misc game fixes/changes

+++ rel/include/OWR.h
@@ -18,7 +18,7 @@ namespace mod::owr
 		int16_t ReplaceGeneralItem(int16_t id, int32_t collection_expr);
 		void OnModuleLoaded(OSModuleInfo* module_info);
 		void SequenceInit();
-		void DrawString(const char* data, int32_t x, int32_t y, uint32_t color, float scale = 1.0);
+		void DrawString(const char* data, float x, float y, uint32_t color, float scale = 1.0f);
 		void LZTest();
 		void NewFileInit();
